Add FieldMathUtilities.java to separate match functions from MapsUtilities's functions
